+++ Assets/LDtkUnity/Editor/ParsedField/Parsed/LDtkParsedPoint.cs
@@ -23,6 +23,12 @@ namespace LDtkUnity.Editor
 
         public object ImportString(object input)
         {
+            if (_process == null)
+            {
+                Debug.LogError("LDtk: Didn't process a field value, field data may be missing");
+                return Vector2.negativeInfinity;
+            }
+            
             //Point can be legally null. for the purposes of the scene drawer, a null point in LDtk will translate to a magic vector2 that tells the scene drawer not to draw
             if (input == null)
             {